opentelemetry (opentelemetry_api v0.5.0)

The types defined here, and referencing records in opentelemetry.hrl are used to store trace information while being collected on the Erlang node.
Thus, while the types are based on protos found in the opentelemetry-proto repo: src/opentelemetry/proto/trace/v1/trace.proto, they are not exact translations because further processing is done after the span has finished and can be vendor specific. For example, there is no count of the number of dropped attributes in the span record. And an attribute's value can be a function to only evaluate the value if it is actually used (at the time of exporting). And the stacktrace is a regular Erlang stack trace.

Types

attribute/0

Specs

attribute() :: {attribute_key(), attribute_value()}

attribute_key/0

Specs

attribute_key() :: :unicode.unicode_binary() | atom()

attribute_value/0

Specs

attribute_value() :: any()

attributes/0

Specs

attributes() :: [attribute()]

event/0

Specs

event() :: record(:event)

event_name/0

Specs

event_name() :: :unicode.unicode_binary() | atom()

events/0

Specs

events() :: [record(:event)]

link/0

Specs

link() :: record(:link)

links/0

Specs

links() :: [record(:link)]

meter/0

Specs

meter() :: {module(), term()}

resource/0

Specs

resource() :: %{
  optional(:unicode.unicode_binary()) => :unicode.unicode_binary()
}

span/0

Specs

span() :: term()

span_ctx/0

Specs

span_ctx() :: record(:span_ctx)

span_id/0

Specs

span_id() :: non_neg_integer()

span_kind/0

Specs

span_kind() :: :internal | :server | :client | :producer | :consumer

span_name/0

Specs

span_name() :: :unicode.unicode_binary() | atom()

status/0

Specs

status() :: record(:status)

status_code/0

Specs

status_code() :: :unset | :ok | :error

text_map/0

Specs

text_map() :: [{:unicode.unicode_binary(), :unicode.unicode_binary()}]

timestamp/0

Specs

timestamp() :: integer()

trace_flags/0

Specs

trace_flags() :: non_neg_integer()

trace_id/0

Specs

trace_id() :: non_neg_integer()

tracer/0

Specs

tracer() :: {module(), term()}

tracestate/0

Specs

tracestate() :: [{:unicode.latin1_chardata(), :unicode.latin1_chardata()}]

Functions

convert_timestamp(Timestamp, Unit)

Specs

convert_timestamp(timestamp(), :erlang.time_unit()) :: integer()

Convert a native monotonic timestamp to POSIX time of any erlang:time_unit(). Meaning the time since Epoch. Epoch is defined to be 00:00:00 UTC, 1970-01-01.

event(Name, Attributes)

Specs

event(name, attributes) :: event() | :undefined
when name: :unicode.unicode_binary(), attributes: attributes()

event(Timestamp, Name, Attributes)

Specs

event(timestamp, name, attributes) :: event() | :undefined
when timestamp: non_neg_integer(),
     name: :unicode.unicode_binary(),
     attributes: attributes()

events(List)

generate_span_id()

Specs

generate_span_id() :: span_id()

Generates a 64 bit random integer to use as a span id.

generate_trace_id()

Specs

generate_trace_id() :: trace_id()

Generates a 128 bit random integer to use as a trace id.

get_meter()

Specs

get_meter() :: meter()

get_meter(Name)

Specs

get_meter(atom()) :: meter()

get_text_map_extractors()

get_text_map_injectors()

get_tracer()

Specs

get_tracer() :: tracer()

get_tracer(Name)

Specs

get_tracer(atom()) :: tracer()

link(SpanCtx)

Specs

link(span_ctx() | :undefined) :: link()

link(Span_ctx, Attributes)

Specs

link(span_ctx() | :undefined, attributes()) :: link()

link(TraceId, SpanId, Attributes, TraceState)

Specs

link(traceId, spanId, attributes, traceState) :: link() | :undefined
when traceId: trace_id(),
     spanId: span_id(),
     attributes: attributes(),
     traceState: tracestate()

link_or_false(TraceId, SpanId, Attributes, TraceState)

links(List)

Specs

links([
  {traceId, spanId, attributes, traceState}
  | span_ctx()
  | {span_ctx(), attributes}
]) :: links()
when traceId: trace_id(),
     spanId: span_id(),
     attributes: attributes(),
     traceState: tracestate()

register_application_meter(Name)

Specs

register_application_meter(atom()) :: boolean()

register_application_tracer(Name)

Specs

register_application_tracer(atom()) :: boolean()

register_meter(Name, Vsn)

Specs

register_meter(atom(), string()) :: boolean()

register_tracer(Name, Vsn)

Specs

register_tracer(atom(), string()) :: boolean()

set_default_meter(Meter)

Specs

set_default_meter(meter()) :: boolean()

set_default_tracer(Tracer)

Specs

set_default_tracer(tracer()) :: boolean()

set_meter(Name, Meter)

Specs

set_meter(atom(), meter()) :: boolean()

set_text_map_extractors(List)

set_text_map_injectors(List)

set_tracer(Name, Tracer)

Specs

set_tracer(atom(), tracer()) :: boolean()

status(Code, Message)

Specs

status(code, message) :: status() | :undefined
when code: status_code(), message: :unicode.unicode_binary()

timestamp()

Specs

timestamp() :: integer()

A monotonically increasing time provided by the Erlang runtime system in the native time unit. This value is the most accurate and precise timestamp available from the Erlang runtime and should be used for finding durations or any timestamp that can be converted to a system time before being sent to another system.

timestamp_to_nano(Timestamp)

Specs

timestamp_to_nano(timestamp()) :: integer()

Convert a native monotonic timestamp to nanosecond POSIX time. Meaning the time since Epoch. Epoch is defined to be 00:00:00 UTC, 1970-01-01.

uniform(X)

verify_and_set_term(Module, TermKey, Behaviour)

Specs

verify_and_set_term(module() | {module(), term()}, term(), atom()) :: boolean()

verify_module_exists(Module)

Specs

verify_module_exists(module() | {module(), term()}) :: boolean()