OpenTelemetry.Ctx (opentelemetry_api v0.5.0)

Ctx is responsible for propagating values within a process that are associated with a particular Trace or set of Baggage. OpenTelemetry.Tracer and OpenTelemetry.Baggage handle updating the Context.

Functions

attach(ctx)

See :otel_ctx.attach/1.

clear()

See :otel_ctx.clear/0.

detach(token)

See :otel_ctx.detach/1.

get_current()

See :otel_ctx.get_current/0.

get_value(key, default)

See :otel_ctx.get_value/2.

get_value(ctx, key, default)

See :otel_ctx.get_value/3.

new()

See :otel_ctx.new/0.

remove(key)

See :otel_ctx.remove/1.

set_value(key, value)

See :otel_ctx.set_value/2.

set_value(ctx, key, value)

See :otel_ctx.set_value/3.