otel_ctx (opentelemetry_api v0.5.0)

Ctx is responsible for propagating values within a process that are associated with a particular Trace or set of Baggage. OpenTelemetry.Tracer and OpenTelemetry.Baggage handle updating the Context.

Types

key/0

Specs

key() :: term()

t/0

Specs

t() :: map()

token/0

Specs

token() :: reference()

value/0

Specs

value() :: term()

Functions

attach(Ctx)

Specs

attach(map()) :: token()

clear()

Specs

clear() :: :ok

detach(Token)

Specs

detach(token()) :: :ok

get_current()

Specs

get_current() :: map()

get_value(Key)

Specs

get_value(term()) :: term()

get_value(Key, Default)

Specs

get_value(term(), term()) :: term()

get_value(Ctx, Key, Default)

Specs

get_value(t(), term(), term()) :: term()

new()

Specs

new() :: t()

remove(Key)

Specs

remove(term()) :: :ok

set_value(Key, Value)

Specs

set_value(term(), term()) :: :ok

set_value(Ctx, Key, Value)

Specs

set_value(t(), term(), term()) :: map()

text_map_extractor(Key, FromText)

text_map_extractor_fun(TextMap, Key, FromText)

text_map_injector(Key, ToText)

text_map_injector_fun(TextMap, Key, ToText)