otel_instrument behaviour (opentelemetry_api v0.5.0)

All measurements are associated with an instrument. To record measurements for an instrument it must first be created with new and then can be referenced by name.

Callbacks

add/2
(optional)

Specs

add(:otel_meter.bound_instrument(), number()) :: :ok

add/4
(optional)

Specs

add(:opentelemetry.meter(), :otel_meter.name(), number(), :otel_meter.labels()) ::
  :ok

definition/1

Specs

definition(:otel_meter.name()) :: :otel_meter.instrument_definition()

definition/2

Specs

definition(:otel_meter.name(), :otel_meter.instrument_opts()) ::
  :otel_meter.instrument_definition()

measurement/2
(optional)

Specs

measurement(:otel_meter.bound_instrument() | :otel_meter.name(), number()) ::
  {:otel_meter.bound_instrument() | :otel_meter.name(), number()}

new/2

Specs

new(:opentelemetry.meter(), :otel_meter.name()) :: boolean()

new/3

Specs

new(:opentelemetry.meter(), :otel_meter.name(), :otel_meter.instrument_opts()) ::
  boolean()

record/2
(optional)

Specs

record(:otel_meter.bound_instrument(), number()) :: :ok

record/4
(optional)

Specs

record(
  :opentelemetry.meter(),
  :otel_meter.name(),
  number(),
  :otel_meter.labels()
) :: :ok