otel_tracer behaviour (opentelemetry_api v1.0.0-rc.2)

Types

traced_fun/1

Specs

traced_fun(T) :: fun((opentelemetry:span_ctx()) -> T).

tracer_ctx/0

Specs

tracer_ctx() :: term().

Functions

add_event(Name, Attributes)

Specs

add_event(Name, Attributes) -> boolean()
             when Name :: opentelemetry:event_name(), Attributes :: opentelemetry:attributes().

add_events(Events)

Specs

add_events(Events) -> boolean() when Events :: opentelemetry:events().

current_span_ctx()

Specs

current_span_ctx() -> opentelemetry:span_ctx() | undefined.

current_span_ctx(Ctx)

Specs

current_span_ctx(otel_ctx:t()) -> opentelemetry:span_ctx() | undefined.

end_span()

Specs

end_span() -> opentelemetry:span_ctx().

non_recording_span(TraceId, SpanId, Traceflags)

Specs

non_recording_span(opentelemetry:trace_id(), opentelemetry:span_id(), opentelemetry:trace_flags()) ->
                      opentelemetry:span_ctx().

Returns a span_ctx record with is_recording set to false. This is mainly for use in propagators when they extract a Span to be used as a parent.

set_attribute(Key, Value)

Specs

set_attribute(Key, Value) -> boolean()
                 when
                     Key :: opentelemetry:attribute_key(),
                     Value :: opentelemetry:attribute_value().

set_attributes(Attributes)

Specs

set_attributes(Attributes) -> boolean() when Attributes :: opentelemetry:attributes().

set_current_span(SpanCtx)

Specs

set_current_span(opentelemetry:span_ctx() | undefined) -> ok.

set_current_span(Ctx, SpanCtx)

Specs

set_current_span(otel_ctx:t(), opentelemetry:span_ctx() | undefined) -> otel_ctx:t() | undefined.

set_status(Status)

Specs

set_status(Status) -> boolean() when Status :: opentelemetry:status().

start_span(Tracer, Name, Opts)

Specs

start_span(opentelemetry:tracer(), opentelemetry:span_name(), otel_span:start_opts()) ->
              opentelemetry:span_ctx().

start_span(Ctx, Tracer, Name, Opts)

Specs

start_span(otel_ctx:t(),
           opentelemetry:tracer(),
           opentelemetry:span_name(),
           otel_span:start_opts()) ->
              opentelemetry:span_ctx().

text_map_propagators(Module)

Specs

text_map_propagators(module()) ->
                        {otel_propagator:text_map_extractor(),
                         otel_propagator:text_map_injector()}.

update_name(Name)

Specs

update_name(Name) -> boolean() when Name :: opentelemetry:span_name().

with_span(Tracer, SpanName, Opts, Fun)

Specs

with_span(opentelemetry:tracer(),
          opentelemetry:span_name(),
          otel_span:start_opts(),
          traced_fun(T)) ->
             T.

with_span(Ctx, Tracer, SpanName, Opts, Fun)

Specs

with_span(otel_ctx:t(),
          opentelemetry:tracer(),
          opentelemetry:span_name(),
          otel_span:start_opts(),
          traced_fun(T)) ->
             T.

Callbacks

start_span/4

Specs

start_span(otel_ctx:t(),
           opentelemetry:tracer(),
           opentelemetry:span_name(),
           otel_span:start_opts()) ->
              opentelemetry:span_ctx().

with_span/5

Specs

with_span(otel_ctx:t(),
          opentelemetry:tracer(),
          opentelemetry:span_name(),
          otel_span:start_opts(),
          traced_fun(T)) ->
             T.